+++ Sources/Model/Atom/AtomFeedEntryLink.swift
@@ -149,7 +149,7 @@ public class AtomFeedEntryLink {
          attribute.
          
          */
-        public var length: Int?
+        public var length: Int64?
         
     }
